Renamed AttemptLogin / BGWorker to ServerManager and moved to Utils
